refactoring: move around event extraction classes, pass 1
The organization of the classes is suboptimal. The events package
contains a bunch of stuff, not necessarily related to each other.

This first pass at refactoring moves the http specific extraction classes
to the endpoint.extractors package.

In the same time, get rid of the EventExtractorUtil interface, which
wasn't really useful.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>