test: add full end-to-end tests
Add a slow test class that brings up the full stack, Jetty included.
This helps us test the public API.

Individual tests pass, not as a whole though due to the shared server
instance. Still trying to figure out if I really need to split this in
multiple classes or if I can make TestNG recycle the server object.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>