endpoints: revisit data flow from the various resources
After analyzing some heap dumps in production, significant overhead was added
by the per-request stats bookeeping in place. This patch revisit the events flow
and centralize stats in EventDeserializerRequestHandler and EventFilterRequestHandler.

It removes EventStats, Stats and SynchronizedTimeWindowStatistics classes in favor of simple
metrics meters.

As part of this refactoring, we can simplify the Guice wiring and get rid of the EventRequestHandlerProvider.
The interfaces EventHandler and ExtractedAnnotation weren't really useful and got removed as well.

Other miscellaneous fixes include:

* Fix the warnings in the various filters by making them explicitly filters of ParsedRequest instances
* Move EventCollector out of the hadoop package into com.ning.metrics.collector.processing as it is
not Hadoop specific
* Move to slf4j wherever possible

Tests:

* Added a bunch of tests for the resources
* Make tests not try to write to /events, this was causing failures reported by Tatu on 16-Dec-2011

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>