metrics: update to 2.1.2
Update Metrics to the 2.1.x API. This is the first step towards updating
the collector to the 0.1.x pom.

Note that we had to update the serialization version as a side effect due
to incompatible changes in the API.

Hadoop has also been updated to the latest 0.20.2x release. We updated the healthcheck
not to rely of FsStatus, which is not always available depending on the version used.
The Hadoop version will be configurable once we update the parent pom.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>